+++ features/product/src/main/java/christian/chamorro/product/domain/models/Product.kt
@@ -5,7 +5,7 @@ data class Product(
     val title: String,
     val condition: String,
     val thumbnail: String,
-    val originalPrice: Long,
+    val originalPrice: Double,
     val initialQuantity: Long,
     val acceptsMercadopago: Boolean,
     val warranty: String,
